remove auth
