+++ sonar-project.properties
@@ -13,7 +13,7 @@ sonar.links.scm=https://github.com/DEFRA/apha-apps-perms-move-animal-ui
 sonar.links.issue=https://github.com/DEFRA/apha-apps-perms-move-animal-ui/issues
 
 sonar.sources=src/
-sonar.exclusions=src/**/*.test.js,src/server/common/test-helpers/*,src/server/common/helpers/auth/*
+sonar.exclusions=src/**/*.test.js,src/server/common/test-helpers/*,src/server/common/helpers/auth/*,src/server/auth/**/*.js
 sonar.tests=src/
 sonar.test.inclusions=src/**/*.test.js
