+++ src/server/auth/callback/index.js
@@ -1,4 +1,5 @@
-import { authCallbackController } from '~/src/server/auth/callback/controller.js'
+import { v4 as uuidv4 } from 'uuid'
+import { addSeconds } from 'date-fns'
 
 const auth = {
   plugin: {
@@ -7,7 +8,35 @@ const auth = {
       server.route({
         method: ['GET', 'POST'],
         path: '/auth/callback',
-        ...authCallbackController
+        options: {
+          auth: 'defra-id'
+        },
+        handler: async (request, h) => {
+          if (request.auth.isAuthenticated) {
+            const { profile } = request.auth.credentials
+            const expiresInSeconds = request.auth.credentials.expiresIn
+            const expiresInMilliSeconds = expiresInSeconds * 1000
+            const expiresAt = addSeconds(new Date(), expiresInSeconds)
+
+            const sessionId = uuidv4()
+            await request.server.app.cache.set(sessionId, {
+              ...profile,
+              isAuthenticated: request.auth.isAuthenticated,
+              token: request.auth.credentials.token,
+              refreshToken: request.auth.credentials.refreshToken,
+              expiresIn: expiresInMilliSeconds,
+              expiresAt
+            })
+
+            request.cookieAuth.set({ sessionId })
+
+            request.logger.info('User has been successfully authenticated')
+          }
+
+          const redirect = request.yar.flash('referrer')?.at(0) ?? '/'
+
+          return h.redirect(redirect)
+        }
       })
     }
   }
